ChatBotAgent/requirements.txt
llama-index==0.10.20
google-generativeai>=0.4.0
python-dotenv
